Add a new metrics indexing behavior
Take job_metrics.json from release periodics and simplify into
a map  where the key is a prometheus metric + optional labels
(foo{bar="baz"}) and the value is a struct with the float value
as a string (like prometheus returns) and an int64 unix timestamp
which was the timestamp of the query value. Write that map to GCS
with a key suitable for time based indexing inside the source
bucket. This index will be used to fetch metrics by job and display
comparisons in ci-search.