functions: allow metrics to have multiple labels
Previously, there was a clause in the conversion function from json to
sqlite that prevented any metric from having more than one label. This
was done for security purposes in order to prevent having big
cardinality explosions. However, this prevented metrics that have valid
use cases from exposing multiple labels such as
`cluster:version:updates:seconds`, which expose the updated version as
well as the version from which the cluster was updated. Thus, to enable
such use cases, we will remove that clause and still make sure to
prevent database explosion by actively reviewing the metrics that are
added to the system.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>