functions: support metrics with multiple labels
The conversion function was working under the assumption that all the
metrics hosted on the bucket were only allowed to have one label. Since
we want to support the use case of metrics with multiple labels such as
`cluster:version:updates:seconds`, the function needed to be updated to
return metrics with multiple labels. To do so, instead of treating the
metric selector as only one label key-value pair, changes were made to
treat it as multiple key-value pairs separated by commas.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>